feat: Changes serialzier gem

+++ app/controllers/consumables_controller.rb
@@ -3,10 +3,11 @@ class ConsumablesController < ApplicationController
   include Searchable
 
   expose :consumables, -> { search(@active_company.consumables.includes_associated, sortable_fields) }
-  expose :consumable
+  expose :consumable, scope: ->{ @active_company.consumables }, find: ->(id, scope){ scope.includes_associated.find(id) }
 
   # GET /consumables
   def index
+    authorize consumables
     paginated_consumables = consumables.page(params[:page] || 1)
 
     render inertia: "Consumables/Index", props: {
@@ -20,6 +21,7 @@ class ConsumablesController < ApplicationController
 
   # GET /consumables/:id
   def show
+    authorize consumable
     render inertia: "Consumables/Show", props: {
       consumable: -> { consumable.render(view: :show) }
     }
@@ -27,18 +29,20 @@ class ConsumablesController < ApplicationController
 
   # GET /consumables/new
   def new
+    authorize Consumable
     render inertia: "Consumables/New", props: {
-      consumable: Consumable.new.render(view: :new),
+      consumable: Consumable.new.render(view: :form_data),
       models: -> { @active_company.models.find_by_category(:Consumable).render },
       vendors: -> { @active_company.vendors.render },
       locations: -> { @active_company.locations.render },
-      manufacturers: -> { @active_company.manufacturers.render(view: :as_options) },
-      categories: -> { @active_company.categories.find_by_type(:item).render(view: :as_options) }
+      manufacturers: -> { @active_company.manufacturers.render(view: :options) },
+      categories: -> { @active_company.categories.find_by_type(:consumable).render(view: :options) }
     }
   end
 
   # GET /consumables/:id/edit
   def edit
+    authorize consumable
     render inertia: "Consumables/Edit", props: {
       consumable: consumable.render(view: :edit),
       models: -> { @active_company.models.find_by_category(:Consumable).render },
@@ -49,6 +53,7 @@ class ConsumablesController < ApplicationController
 
   # GET /consumables/:id/checkout
   def checkout
+    authorize consumable
     redirect_to consumable if consumable.qty == 0
 
     assignment = Assignment.new
@@ -57,15 +62,16 @@ class ConsumablesController < ApplicationController
 
     render inertia: "Consumables/Checkout", props: {
       consumable: consumable.render,
-      assignment: assignment.render(view: :new),
-      items: -> { @active_company.items.select([:id, :name, :default_location_id]).render(view: :as_options) },
-      people: -> { @active_company.people.select([:id, :first_name, :last_name, :location_id]).render(view: :as_options) },
-      locations: -> { @active_company.locations.select([:id, :slug, :name]).render(view: :as_options) },
+      assignment: assignment.render(view: :form_data),
+      items: -> { @active_company.items.select([:id, :name, :default_location_id]).render(view: :options) },
+      people: -> { @active_company.people.select([:id, :first_name, :last_name, :location_id]).render(view: :options) },
+      locations: -> { @active_company.locations.select([:id, :slug, :name]).render(view: :options) },
     }
   end
 
   # GET /consumables/:id/checkin
   def checkin
+    authorize consumable
     redirect_to consumable if consumable.assignments.empty?
     assignment = consumable.assignment
     assignment.returned_at = Time.current
@@ -74,13 +80,14 @@ class ConsumablesController < ApplicationController
     render inertia: "Consumables/Checkin", props: {
       consumable: consumable.render,
       assignment: assignment.render,
-      locations: -> { @active_company.locations.select([:id, :slug, :name]).render(view: :as_options) },
+      locations: -> { @active_company.locations.select([:id, :slug, :name]).render(view: :options) },
       statuses: -> { StatusLabel.all.render }
     }
   end
 
   # POST /consumables
   def create
+    authorize Consumable
     consumable.company = @active_company
     if consumable.save
       redirect_to consumable, notice: 'Consumable was successfully created'
@@ -91,6 +98,7 @@ class ConsumablesController < ApplicationController
 
   # PATCH/PUT /consumables/:id
   def update
+    authorize consumable
     if consumable.update(consumable_params)
       redirect_to consumable, notice: 'Consumable was successfully updated'
     else
@@ -100,6 +108,7 @@ class ConsumablesController < ApplicationController
 
   # DELETE /consumables/:id
   def destroy
+    authorize consumable
     consumable.destroy
     respond_to do |format|
       format.html { redirect_to consumables_url, notice: 'Consumable was successfully destroyed.' }
@@ -114,6 +123,6 @@ class ConsumablesController < ApplicationController
   end
 
   def consumable_params
-    params.require(:consumable).permit(:name, :min_qty, :qty, :cost, :requestable, :notes, :manufacturer_id, :category_id, :vendor_id, :default_location_id)
+    params.require(:consumable).permit(:name, :min_qty, :qty, :cost, :serial, :asset_tag, :cost_currency, :requestable, :notes, :model_id, :manufacturer_id, :category_id, :vendor_id, :default_location_id)
   end
 end
