+++ app/controllers/departments_controller.rb
@@ -3,10 +3,11 @@ class DepartmentsController < ApplicationController
   include ContactableConcern
 
   expose :departments, -> { search(@active_company.departments.includes_associated, sortable_fields) }
-  expose :department, -> { @active_company.departments.find_by_slug(params[:slug]) }
+  expose :department, id: ->{ params[:slug] }, scope: ->{ @active_company.departments.includes_associated }, find_by: :slug
 
   # GET /departments
   def index
+    authorize departments
     paginated_departments = departments.page(params[:page] || 1)
 
     render inertia: "Departments/Index", props: {
@@ -20,12 +21,13 @@ class DepartmentsController < ApplicationController
 
   # GET /departments/:slug
   def show
+    authorize department
     render inertia: "Departments/Show", props: {
       department: department.render(view: :show),
       items: InertiaRails.lazy(-> {
         paginated_items = department.items.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_items.render(view: :associations),
+          data: paginated_items.render,
           pagination: {
             count: department.items.size,
             **pagination_data(paginated_items)
@@ -35,7 +37,7 @@ class DepartmentsController < ApplicationController
       accessories: InertiaRails.lazy(-> {
         paginated_accessories = department.accessories.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_accessories.render(view: :associations),
+          data: paginated_accessories.render,
           pagination: {
             count: department.accessories.size,
             **pagination_data(paginated_accessories)
@@ -45,7 +47,7 @@ class DepartmentsController < ApplicationController
       consumables: InertiaRails.lazy(-> {
         paginated_consumables = department.consumables.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_consumables.render(view: :associations),
+          data: paginated_consumables.render,
           pagination: {
             count: department.consumables.size,
             **pagination_data(paginated_consumables)
@@ -55,7 +57,7 @@ class DepartmentsController < ApplicationController
       components: InertiaRails.lazy(-> {
         paginated_components = department.components.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_components.render(view: :associations),
+          data: paginated_components.render,
           pagination: {
             count: department.components.size,
             **pagination_data(paginated_components)
@@ -65,7 +67,7 @@ class DepartmentsController < ApplicationController
       licenses: InertiaRails.lazy(-> {
         paginated_licenses = department.licenses.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_licenses.render(view: :associations),
+          data: paginated_licenses.render,
           pagination: {
             count: department.licenses.size,
             **pagination_data(paginated_licenses)
@@ -75,7 +77,7 @@ class DepartmentsController < ApplicationController
       people: InertiaRails.lazy(-> {
         paginated_people = department.people.includes_associated.page(params[:page] || 1)
         {
-          data: paginated_people.render(view: :associations),
+          data: paginated_people.render,
           pagination: {
             count: department.people.size,
             **pagination_data(paginated_people)
@@ -87,22 +89,25 @@ class DepartmentsController < ApplicationController
 
   # GET /departments/new
   def new
+    authorize Department
     render inertia: "Departments/New", props: {
-      department: Department.new.render(view: :new),
-      locations: -> { @active_company.locations.render(view: :as_options) }
+      department: Department.new.render(view: :form_data),
+      locations: -> { @active_company.locations.render(view: :options) }
     }
   end
 
   # GET /departments/:slug/edit
   def edit
+    authorize department
     render inertia: "Departments/Edit", props: {
       department: department.render(view: :edit),
-      locations: -> { @active_company.locations.render(view: :as_options) }
+      locations: -> { @active_company.locations.render(view: :options) }
     }
   end
 
   # POST /departments
   def create
+    authorize Department
     department = Department.new(department_params)
     department.company = @active_company
 
@@ -115,6 +120,7 @@ class DepartmentsController < ApplicationController
 
   # PATCH/PUT /departments/:slug
   def update
+    authorize department
     if department.update(department_params)
       redirect_to department, notice: 'Department was successfully updated'
     else
@@ -124,6 +130,7 @@ class DepartmentsController < ApplicationController
 
   # DELETE /departments/:slug
   def destroy
+    authorize department
     department.destroy
     redirect_to departments_url, notice: 'Department was successfully destroyed.'
   end
